added more

+++ Core/entities/material.cpp
@@ -101,7 +101,7 @@ namespace Zion
 		std::string str1 = std::string("isTexture[") + std::to_string(index) + std::string("]");
 		std::string str2 = std::string("texColor[") + std::to_string(index) + std::string("]");
 
-		shader.setUniform1i((GLchar *)str.c_str(), index);
+		shader.setUniform1i((GLchar *)str.c_str(), (GLint)index);
 		shader.setUniform1i((GLchar *)str1.c_str(), (int)mat.texure.isLoaded());
 		shader.setUniform4f((GLchar *)str2.c_str(), mat.base_color);
 	}
